Refactored LoginPage
