+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 
 import { Home } from "./components/home/Home";
-import { Login } from "./components/login/Login";
+import { LoginPage } from "./components/login/LoginPage";
 import { Header } from "./components/pageElements/header/Header";
 
 import "./app.scss";
@@ -23,7 +23,7 @@ const App: React.FunctionComponent = () => {
 							<h1>Feed</h1>
 						</Route>
 						<Route path="/login">
-							<Login />
+							<LoginPage />
 						</Route>
 					</Switch>
