+++ src/components/login/UserDetailsForm/detailsValidator.scss
