+++ src/components/login/UserDetailsForm/validators/validators.ts
@@ -1,5 +1,4 @@
 
-
 const validators = {
 
 	numberRegex: /\d/g,
